Create blackcomedians
A society’s comedians tell exactly where the social lines are, and nowhere are this more apparent than with Black Comedian. They have actually done a lot of cutting-edge work on black cultural pathology.  In the words of Chappelle doing his rap-comedian act, “I got jokes and jokes and jokes and jokes…” Comedy as truth telling — “it’s funny because it’s true” — and it starts here.  Visit http://www.youtube.com/Blackcomediansuncut